+++ monitoring/monitoring-api/src/main/java/com/rbmhtechnology/vind/monitoring/model/request/RootTimeMixIn.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2018 Redlink GmbH.
+ */
+package com.rbmhtechnology.vind.monitoring.model.request;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+/**
+ * Created on 04.04.18.
+ */
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class RootTimeMixIn {
+}