+++ monitoring/monitoring-api/src/main/java/com/rbmhtechnology/vind/monitoring/model/request/facet/FacetMixin.java
@@ -15,12 +15,10 @@ import com.rbmhtechnology.vind.api.query.facet.Facet;
 
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 @JsonPropertyOrder({"type"})
-@JsonIgnoreProperties(ignoreUnknown = true, value = {"fieldDescriptor","descriptor", "log","tagedPivots","fieldName" })
-
+@JsonIgnoreProperties(ignoreUnknown = true, value = {"fieldDescriptor","descriptor", "log","tagedPivots","name" })
 public abstract class FacetMixin extends Facet {
 
     @JsonProperty("field")
-    @Override
-    public abstract String getName();
+    public abstract String getFieldName();
 
 }